Routing Rework
Users can only access the pages they are meant to access. Moved the login component to be a page because programmatically navigating there throws a 404... The same happens sometimes when a user logs in. This commit is to test the navigate() in a production environment. onAuthStateChanged is still in the layout, but is much simpler.